+++ src/debug.rs
@@ -36,6 +36,7 @@ impl Plugin for DebugPlugin {
                 .register_inspectable::<EnemyType>()
                 .register_inspectable::<Patrol>()
                 .register_inspectable::<Health>()
+                // .register_inspectable::<PlayerAnimationState>()
                 .add_stage_after(CoreStage::Update, DEBUG, SystemStage::single_threaded())
                 .add_system_to_stage(DEBUG, debug_collisions)
                 .add_system_to_stage(DEBUG, update_debug_collisions)